+++ packages/producer/src/services/render/stages/captureHdrStage.ts
@@ -57,17 +57,14 @@ import type { ProducerLogger } from "../../../logger.js";
 import { createHdrImageTransferCache } from "../../hdrImageTransferCache.js";
 import {
   type HdrCompositeContext,
-  type HdrDiagnostics,
-  type HdrPerfCollector,
   type HdrTransitionMeta,
   type HdrVideoFrameSource,
-  type ProgressCallback,
-  type RenderJob,
   type TransitionRange,
   closeHdrVideoFrameSource,
-  createHdrPerfCollector,
   resolveCompositeTransfer,
-} from "../../renderOrchestrator.js";
+} from "../../hdrCompositor.js";
+import { type HdrPerfCollector, createHdrPerfCollector } from "../hdrPerf.js";
+import type { HdrDiagnostics, ProgressCallback, RenderJob } from "../../renderOrchestrator.js";
 import type { CompositionMetadata } from "../shared.js";
 import {
   decodeHdrImageBuffers,
